os/wqueue : Integrate user work queue and kernel work queue
User work queue and kernel work queue had the same operation.
So i've integrated the code into one. Builds with kbin and ubin, respectively.
