+++ lib/libc/Kconfig
@@ -455,39 +455,3 @@ config LIBC_EXECFUNCS
 		non-standard binfmt function 'exec', and then (2) exit(0).  Since
 		the new thread will be terminated by the exec[l|v] call, it really
 		served no purpose other than to suport Unix compatility.
-
-comment "Non-standard Library Support"
-
-if BUILD_PROTECTED || BUILD_KERNEL
-
-config LIB_USRWORK
-	bool "User mode worker thread"
-	default y
-	depends on !DISABLE_SIGNALS
-	---help---
-		User space work queues can also be made available for deferred
-		processing in the TinyAra kernel build.
-
-if LIB_USRWORK
-
-config LIB_USRWORKPRIORITY
-	int "User mode priority worker thread priority"
-	default 100
-	---help---
-		The execution priority of the user-mode priority worker thread.  Default: 100
-
-config LIB_USRWORKPERIOD
-	int "User mode worker thread period"
-	default 100000
-	---help---
-		How often the lower priority worker thread checks for work in units
-		of microseconds. Default: 100*1000 (100 MS).
-
-config LIB_USRWORKSTACKSIZE
-	int "User mode worker thread stack size"
-	default 2048
-	---help---
-		The stack size allocated for the lower priority worker thread.  Default: 2K.
-
-endif # LIB_USRWORK
-endif # BUILD_PROTECTED || BUILD_KERNEL
