+++ os/Directories.mk
@@ -148,6 +148,7 @@ endif
 CONTEXTDIRS = $(APPDIR)
 CONTEXTDIRS += $(TOOLSDIR)
 CONTEXTDIRS += mm
+CONTEXTDIRS += wqueue
 #ifeq ($(CONFIG_ENABLE_IOTIVITY),y)
 #CONTEXTDIRS += $(EXTDIR)$(DELIM)iotivity
 #endif
@@ -156,7 +157,7 @@ OTHERDIRS = $(LIBRARIES_DIR)
 
 ifeq ($(CONFIG_BUILD_PROTECTED),y)
 
-USERDIRS += $(LIB_DIR)$(DELIM)libc mm $(USER_ADDONS)
+USERDIRS += $(LIB_DIR)$(DELIM)libc mm wqueue$(USER_ADDONS)
 ifeq ($(CONFIG_HAVE_CXX),y)
 USERDIRS += $(LIB_DIR)$(DELIM)libxx
 endif
@@ -164,14 +165,14 @@ endif
 else
 ifeq ($(CONFIG_BUILD_KERNEL),y)
 
-USERDIRS += $(LIB_DIR)$(DELIM)libc mm
+USERDIRS += $(LIB_DIR)$(DELIM)libc mm wqueue
 ifeq ($(CONFIG_HAVE_CXX),y)
 USERDIRS += $(LIB_DIR)$(DELIM)libxx
 endif
 
 else
 
-NONFSDIRS += $(LIB_DIR)$(DELIM)libc mm
+NONFSDIRS += $(LIB_DIR)$(DELIM)libc mm wqueue
 OTHERDIRS += $(USER_ADDONS)
 ifeq ($(CONFIG_HAVE_CXX),y)
 NONFSDIRS += $(LIB_DIR)$(DELIM)libxx
