+++ os/wqueue/.gitignore
@@ -1,0 +1,5 @@
+/Make_bin.dep
+/Make_ubin.dep
+/Make_kbin.dep
+/.depend
+/*.lib
