+++ os/wqueue/Make.defs
@@ -16,9 +16,9 @@
 #
 ###########################################################################
 ############################################################################
-# kernel/Makefile
+# wqueue/Make.defs
 #
-#   Copyright (C) 2007-2014 Gregory Nutt. All rights reserved.
+#   Copyright (C) 2014 Gregory Nutt. All rights reserved.
 #   Author: Gregory Nutt <gnutt@nuttx.org>
 #
 # Redistribution and use in source and binary forms, with or without
@@ -50,65 +50,15 @@
 #
 ############################################################################
 
--include $(TOPDIR)/Make.defs
+# Add work queue files
 
-ASRCS =
-CSRCS =
+ifeq ($(CONFIG_SCHED_WORKQUEUE),y)
 
-VPATH =
-DEPPATH = --dep-path .
+CSRCS += work_queue.c work_process.c work_cancel.c work_signal.c
 
-include init/Make.defs
-include irq/Make.defs
-include paging/Make.defs
-include group/Make.defs
-include sched/Make.defs
-include task/Make.defs
-include errno/Make.defs
-include wdog/Make.defs
-include semaphore/Make.defs
-include binary_manager/Make.defs
-include signal/Make.defs
-include pthread/Make.defs
-include messaging/Make.defs
-include mqueue/Make.defs
-include clock/Make.defs
-include timer/Make.defs
-include environ/Make.defs
-include debug/Make.defs
+# Include wqueue build support
 
-AOBJS = $(ASRCS:.S=$(OBJEXT))
-COBJS = $(CSRCS:.c=$(OBJEXT))
+DEPPATH += --dep-path wqueue
+VPATH += :wqueue
 
-SRCS = $(ASRCS) $(CSRCS)
-OBJS = $(AOBJS) $(COBJS)
-
-BIN = libkernel$(LIBEXT)
-
-all: $(BIN)
-.PHONY: context depend clean distclean
-
-$(AOBJS): %$(OBJEXT): %.S
-	$(call ASSEMBLE, $<, $@)
-
-$(COBJS): %$(OBJEXT): %.c
-	$(call COMPILE, $<, $@)
-
-$(BIN): $(OBJS)
-	$(call ARCHIVE, $@, $(OBJS))
-
-.depend: Makefile $(SRCS)
-	$(Q) $(MKDEP) $(DEPPATH) "$(CC)" -- $(CFLAGS) -- $(SRCS) >Make.dep
-	$(Q) touch $@
-
-depend: .depend
-
-clean:
-	$(call DELFILE, $(BIN))
-	$(call CLEAN)
-
-distclean: clean
-	$(call DELFILE, Make.dep)
-	$(call DELFILE, .depend)
-
--include Make.dep
+endif # CONFIG_SCHED_WORKQUEUE
