+++ os/wqueue/kbin/.gitignore
@@ -1,0 +1,8 @@
+/*.asm
+/*.obj
+/*.rel
+/*.lst
+/*.sym
+/*.adb
+/*.lib
+/*.src
