+++ os/wqueue/uwqueue/uwork_lock.c
@@ -16,9 +16,9 @@
  *
  ****************************************************************************/
 /****************************************************************************
- * libc/wqueue/work_cancel.c
+ * wqueue/uwqueue/uwork_lock.c
  *
- *   Copyright (C) 2009-2010, 2012-2014 Gregory Nutt. All rights reserved.
+ *   Copyright (C) 2014 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
  *
  * Redistribution and use in source and binary forms, with or without
@@ -56,14 +56,12 @@
 
 #include <tinyara/config.h>
 
-#include <queue.h>
+#include <pthread.h>
+#include <semaphore.h>
 #include <assert.h>
 #include <errno.h>
 
-#include <tinyara/arch.h>
-#include <tinyara/wqueue.h>
-
-#include "wqueue/wqueue.h"
+#include "wqueue.h"
 
 #if defined(CONFIG_LIB_USRWORK) && !defined(__KERNEL__)
 
@@ -88,91 +86,66 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Name: work_qcancel
- *
- * Description:
- *   Cancel previously queued work.  This removes work from the work queue.
- *   After work has been cancelled, it may be re-queue by calling work_queue()
- *   again.
- *
- * Input parameters:
- *   qid    - The work queue ID
- *   work   - The previously queue work structure to cancel
- *
- * Returned Value:
- *   Zero (OK) on success, a negated errno on failure.  This error may be
- *   reported:
- *
- *   -ENOENT - There is no such work queued.
- *   -EINVAL - An invalid work queue was specified
- *
- ****************************************************************************/
-
-static int work_qcancel(FAR struct usr_wqueue_s *wqueue, FAR struct work_s *work)
-{
-	int ret = -ENOENT;
-
-	DEBUGASSERT(work != NULL);
-
-	/* Get exclusive access to the work queue */
-
-	while (work_lock() < 0);
-
-	/* Cancelling the work is simply a matter of removing the work structure
-	 * from the work queue.  This must be done with interrupts disabled because
-	 * new work is typically added to the work queue from interrupt handlers.
-	 */
-
-	if (work->worker != NULL) {
-		/* A little test of the integrity of the work queue */
-
-		DEBUGASSERT(work->dq.flink || (FAR dq_entry_t *)work == wqueue->q.tail);
-		DEBUGASSERT(work->dq.blink || (FAR dq_entry_t *)work == wqueue->q.head);
-
-		/* Remove the entry from the work queue and make sure that it is
-		 * mark as available (i.e., the worker field is nullified).
-		 */
-
-		dq_rem((FAR dq_entry_t *)work, &wqueue->q);
-		work->worker = NULL;
-		ret = OK;
-	}
-
-	work_unlock();
-	return ret;
-}
-
-/****************************************************************************
  * Public Functions
  ****************************************************************************/
 
 /****************************************************************************
- * Name: work_cancel
+ * Name: work_lock
  *
  * Description:
- *   Cancel previously queued user-mode work.  This removes work from the
- *   user mode work queue.  After work has been cancelled, it may be re-queue
- *   by calling work_queue() again.
+ *   Lock the user-mode work queue.
  *
  * Input parameters:
- *   qid    - The work queue ID (must be USRWORK)
- *   work   - The previously queue work structure to cancel
+ *   None
  *
  * Returned Value:
  *   Zero (OK) on success, a negated errno on failure.  This error may be
  *   reported:
  *
- *   -ENOENT - There is no such work queued.
+ *   -EINTR - Wait was interrupted by a signal
  *
  ****************************************************************************/
 
-int work_cancel(int qid, FAR struct work_s *work)
+int work_lock(void)
 {
-	if (qid == USRWORK) {
-		return work_qcancel(&g_usrwork, work);
-	} else {
-		return -EINVAL;
+	int ret;
+#ifdef CONFIG_BUILD_PROTECTED
+	ret = sem_wait(&g_usrsem);
+	if (ret < 0) {
+		DEBUGASSERT(errno == EINTR);
+		return -EINTR;
 	}
+#else
+	ret = pthread_mutex_lock(&g_usrmutex);
+	if (ret != 0) {
+		DEBUGASSERT(ret == EINTR);
+		return -EINTR;
+	}
+#endif
+	return ret;
+}
+
+/****************************************************************************
+ * Name: work_unlock
+ *
+ * Description:
+ *   Unlock the user-mode work queue.
+ *
+ * Input parameters:
+ *   None
+ *
+ * Returned Value:
+ *   None
+ *
+ ****************************************************************************/
+
+void work_unlock(void)
+{
+#ifdef CONFIG_BUILD_PROTECTED
+	(void)sem_post(&g_usrsem);
+#else
+	(void)pthread_mutex_unlock(&g_usrmutex);
+#endif
 }
 
 #endif							/* CONFIG_LIB_USRWORK && !__KERNEL__ */
