+++ os/wqueue/uwqueue/uwork_thread.c
@@ -16,9 +16,9 @@
  *
  ****************************************************************************/
 /****************************************************************************
- * kernel/wqueue/kwork_cancel.c
+ * wqueue/uwqueue/uwork_usrthread.c
  *
- *   Copyright (C) 2014 Gregory Nutt. All rights reserved.
+ *   Copyright (C) 2009-2014 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
  *
  * Redistribution and use in source and binary forms, with or without
@@ -56,150 +56,190 @@
 
 #include <tinyara/config.h>
 
-#include <queue.h>
-#include <assert.h>
+#include <sys/types.h>
+#include <stdint.h>
+#include <unistd.h>
+#include <pthread.h>
+#include <sched.h>
 #include <errno.h>
+#include <assert.h>
+#include <queue.h>
 
-#include <tinyara/arch.h>
 #include <tinyara/wqueue.h>
+#include <tinyara/clock.h>
 
-#include "wqueue/wqueue.h"
+#include "wqueue.h"
 
-#ifdef CONFIG_SCHED_WORKQUEUE
+#if defined(CONFIG_LIB_USRWORK) && !defined(__KERNEL__)
 
 /****************************************************************************
  * Pre-processor Definitions
  ****************************************************************************/
+
+/* Use CLOCK_MONOTONIC if it is available.  CLOCK_REALTIME can cause bad
+ * delays if the time is changed.
+ */
+
+#ifdef CONFIG_CLOCK_MONOTONIC
+#define WORK_CLOCK CLOCK_MONOTONIC
+#else
+#define WORK_CLOCK CLOCK_REALTIME
+#endif
+
+#ifndef MIN
+#define MIN(a, b) ((a) < (b) ? (a) : (b))
+#endif
 
 /****************************************************************************
  * Private Type Declarations
  ****************************************************************************/
 
 /****************************************************************************
- * Public Variables
+ * Public Data
  ****************************************************************************/
 
+/* The state of the user mode work queue. */
+
+struct wqueue_s g_usrwork;
+
+/* This semaphore supports exclusive access to the user-mode work queue */
+
+#ifdef CONFIG_BUILD_PROTECTED
+sem_t g_usrsem;
+#else
+pthread_mutex_t g_usrmutex;
+#endif
+
 /****************************************************************************
- * Private Variables
+ * Private Data
  ****************************************************************************/
 
 /****************************************************************************
  * Private Functions
  ****************************************************************************/
 
-#if defined(CONFIG_SCHED_HPWORK) || defined(CONFIG_SCHED_LPWORK)
 /****************************************************************************
- * Name: work_qcancel
+ * Name: work_usrthread
  *
  * Description:
- *   Cancel previously queued work.  This removes work from the work queue.
- *   After work has been cancelled, it may be re-queue by calling work_queue()
- *   again.
+ *   This is the worker thread that performs the actions placed on the user
+ *   work queue.
+ *
+ *   This is a user mode work queue.  It must be used by applications for
+ *   miscellaneous operations.  The user work thread must be started by
+ *   application start-up logic by calling work_usrstart().
  *
  * Input parameters:
- *   qid    - The work queue ID
- *   work   - The previously queue work structure to cancel
+ *   argc, argv (not used)
  *
  * Returned Value:
- *   Zero (OK) on success, a negated errno on failure.  This error may be
- *   reported:
- *
- *   -ENOENT - There is no such work queued.
- *   -EINVAL - An invalid work queue was specified
+ *   Does not return
  *
  ****************************************************************************/
 
-static int work_qcancel(FAR struct kwork_wqueue_s *wqueue, FAR struct work_s *work)
+#ifdef CONFIG_BUILD_PROTECTED
+static int work_usrthread(int argc, char *argv[])
+#else
+static pthread_addr_t work_usrthread(pthread_addr_t arg)
+#endif
 {
-	struct work_s *cur_work;
-	irqstate_t flags;
-	int ret = -ENOENT;
+	/* Loop forever */
 
-	DEBUGASSERT(work != NULL);
-
-	/* Cancelling the work is simply a matter of removing the work structure
-	 * from the work queue.  This must be done with interrupts disabled because
-	 * new work is typically added to the work queue from interrupt handlers.
-	 */
-
-	flags = irqsave();
-	if (work->worker != NULL) {
-		/* A little test of the integrity of the work queue */
-
-		DEBUGASSERT(work->dq.flink || (FAR dq_entry_t *)work == wqueue->q.tail);
-		DEBUGASSERT(work->dq.blink || (FAR dq_entry_t *)work == wqueue->q.head);
-
-		/* check whether requested work is in queue list or not */
-		cur_work = (struct work_s *)wqueue->q.head;
-		do {
-			if (cur_work == NULL) {
-				irqrestore(flags);
-				return -ENOENT;
-			} else if (cur_work == work) {
-				break;
-			}
-
-			cur_work = (struct work_s *)cur_work->dq.flink;
-		} while (1);
-
-		/* Remove the entry from the work queue and make sure that it is
-		 * mark as available (i.e., the worker field is nullified).
+	for (;;) {
+		/* Then process queued work.  We need to keep the work queue locked
+		 * while we process items in the work list.
 		 */
 
-		dq_rem((FAR dq_entry_t *)work, &wqueue->q);
-		work->worker = NULL;
-		ret = OK;
+		work_process(&g_usrwork, 0);
 	}
 
-	irqrestore(flags);
-	return ret;
-}
+#ifdef CONFIG_BUILD_PROTECTED
+	return OK;					/* To keep some compilers happy */
+#else
+	return NULL;				/* To keep some compilers happy */
 #endif
+}
 
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
 
 /****************************************************************************
- * Name: work_cancel
+ * Name: work_usrstart
  *
  * Description:
- *   Cancel previously queued user-mode work.  This removes work from the
- *   user mode work queue.  After work has been cancelled, it may be re-queue
- *   by calling work_queue() again.
+ *   Start the user mode work queue.
  *
  * Input parameters:
- *   qid    - The work queue ID (must be HPWORK or LPWORK)
- *   work   - The previously queue work structure to cancel
+ *   None
  *
  * Returned Value:
- *   Zero (OK) on success, a negated errno on failure.  This error may be
- *   reported:
- *
- *   -ENOENT - There is no such work queued.
- *   -EINVAL - An invalid work queue was specified
+ *   The task ID of the worker thread is returned on success.  A negated
+ *   errno value is returned on failure.
  *
  ****************************************************************************/
 
-int work_cancel(int qid, FAR struct work_s *work)
+int work_usrstart(void)
 {
-#ifdef CONFIG_SCHED_HPWORK
-	if (qid == HPWORK) {
-		/* Cancel high priority work */
+	/* Initialize work queue data structures */
 
-		return work_qcancel((FAR struct kwork_wqueue_s *)&g_hpwork, work);
-	} else
-#endif
-#ifdef CONFIG_SCHED_LPWORK
-		if (qid == LPWORK) {
-			/* Cancel low priority work */
+	g_usrwork.delay = CONFIG_LIB_USRWORKPERIOD / USEC_PER_TICK;
+	dq_init(&g_usrwork.q);
 
-			return work_qcancel((FAR struct kwork_wqueue_s *)&g_lpwork, work);
-		} else
-#endif
-		{
-			return -EINVAL;
+#ifdef CONFIG_BUILD_PROTECTED
+	{
+		/* Set up the work queue lock */
+
+		(void)sem_init(&g_usrsem, 0, 1);
+
+		/* Start a user-mode worker thread for use by applications. */
+
+		g_usrwork.worker[0].pid = task_create("uwork", CONFIG_LIB_USRWORKPRIORITY, CONFIG_LIB_USRWORKSTACKSIZE, (main_t)work_usrthread, (FAR char *const *)NULL);
+
+		DEBUGASSERT(g_usrwork.worker[0].pid > 0);
+		if (g_usrwork.worker[0].pid < 0) {
+			int errcode = errno;
+			DEBUGASSERT(errcode > 0);
+			return -errcode;
 		}
+		g_usrwork.worker[0].busy = true;
+		return g_usrwork.worker[0].pid;
+	}
+#else
+	{
+		pthread_t usrwork;
+		pthread_attr_t attr;
+		struct sched_param param;
+		int status;
+
+		/* Set up the work queue lock */
+
+		(void)pthread_mutex_init(&g_usrmutex, NULL);
+
+		/* Start a user-mode worker thread for use by applications. */
+
+		(void)pthread_attr_init(&attr);
+		(void)pthread_attr_setstacksize(&attr, CONFIG_LIB_USRWORKSTACKSIZE);
+
+		param.sched_priority = CONFIG_LIB_USRWORKPRIORITY;
+		(void)pthread_attr_setschedparam(&attr, &param);
+
+		status = pthread_create(&usrwork, &attr, work_usrthread, NULL);
+		if (status != 0) {
+			return -status;
+		}
+
+		/* Detach because the return value and completion status will not be
+		 * requested.
+		 */
+
+		(void)pthread_detach(usrwork);
+
+		g_usrwork.worker[0].pid = (pid_t)usrwork;
+		g_usrwork.worker[0].busy = true;
+		return g_usrwork.worker[0].pid;
+	}
+#endif
 }
 
-#endif							/* CONFIG_SCHED_WORKQUEUE */
+#endif							/* CONFIG_LIB_USRWORK && !__KERNEL__ */
